server/devices/serializers.py
from rest_framework import serializers
from .models import Device


class DeviceRegistrationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Device
        fields = ["name"]


class DeviceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Device
        fields = ["id", "name", "created_at", "last_seen_at"]
