server/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-api-key>=3.0
psycopg2-binary>=2.9
django-storages>=1.14
boto3>=1.34
gunicorn>=21.0
